+++ src/settings/NoticePolicy/utils/notice-description.js
@@ -20,6 +20,7 @@ export default (eventId) => {
     loanUserInitiatedEventsIds.RENEWED,
     loanUserInitiatedEventsIds.MANUAL_DUE_DATE_CHANGE,
     loanUserInitiatedEventsIds.ITEM_RECALLED,
+    loanUserInitiatedEventsIds.HOLD_REQUEST_FOR_ITEM,
     requestItemStateChangeEventsIds.AVAILABLE,
     ...Object.values(requestUserInitiatedEventsIds),
     ...Object.values(requestTimeBasedEventsIds),