+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Hide button when there are no permission for it. Refs UICIRC-1130.
 * User gets an error when try to fill in Lost item processing fee field with > 0. Refs UICIRC-1249.
 * Fixed a problem with clearing field values. Refs UICIRC-1287.
+* Added a new triggering event for hold notices. Refs UICIRC-1245.
 
 ## [11.0.2](https://github.com/folio-org/ui-circulation/tree/v11.0.2) (2024-10-07)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v11.0.1...v11.0.2)